requirements.txt
gunicorn
fastapi
uvicorn
requests
python-dotenv
